LayoutTests/platform/gtk/imported/w3c/web-platform-tests/css/selectors/focus-visible-027-expected.txt



FAIL CSS Test (Selectors): :focus-visible after click and input type change assert_equals: Type change to text might cause :focus-visible to start matching expected false but got true
